fix(c): C-02 step 3 — extract requireAdvisorSession helper, refactor 3 routes
Why: 14 advisor-auth routes each contained an identical ~24-line
getAdvisorId() closure that duplicated the dual-auth resolution logic
(Supabase Auth JWT + legacy advisor_session cookie). Duplication meant
any future auth change had to be applied in 14 places — a maintenance
and correctness risk flagged in audit §4 (admin scope / callgraph).

This step extracts that logic into lib/require-advisor-session.ts and
refactors notifications/route.ts, request-review/route.ts, and
topup/route.ts to import and call it. Admin client is retained inside
the helper because: (1) the "Public can view active professionals" SELECT
policy only covers status='active', but advisor lookups need 'pending'
too; (2) the "Advisor can view own profile" policy only matches
auth.uid() = auth_user_id, not the email fallback we need; (3) legacy
cookie sessions have no auth.uid() linkage at all.

Verified callers: all three routes are advisor-portal-only (no public or
anon paths). Both GET and PATCH on notifications, POST on request-review,
POST and GET on topup all go through requireAdvisorSession → admin
professionals lookup. No change to business logic; auth resolution is
identical to the prior getAdvisorId() closures.

Idempotency: pure refactor — no schema changes, no behaviour change.
Rolling back this commit restores the prior inline getAdvisorId()
implementations.

Refs: #216
Audit: docs/audits/codebase-health-2026-04-24.md §4 (admin scope)
Queue: docs/audits/REMEDIATION_QUEUE.md C-02
