+++ app/api/advisor-auth/topup/route.ts
@@ -1,36 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
-import { createClient } from "@/lib/supabase/server";
 import { createAdminClient } from "@/lib/supabase/admin";
 import { getStripe } from "@/lib/stripe";
 import { getSiteUrl } from "@/lib/url";
 import { DEFAULT_TOPUP_CENTS } from "@/lib/advisor-billing";
 import { isRateLimited } from "@/lib/rate-limit";
-
-async function getAdvisorId(request: NextRequest): Promise<number | null> {
-  const supabase = await createClient();
-  const admin = createAdminClient();
-  
-  const { data: { user } } = await supabase.auth.getUser();
-  if (user) {
-    const { data: advisor } = await admin
-      .from("professionals")
-      .select("id")
-      .or(`auth_user_id.eq.${user.id},email.eq.${user.email}`)
-      .in("status", ["active", "pending"])
-      .maybeSingle();
-    if (advisor) return advisor.id;
-  }
-  
-  const sessionToken = request.cookies.get("advisor_session")?.value;
-  if (!sessionToken) return null;
-  const { data } = await admin
-    .from("advisor_sessions")
-    .select("professional_id, expires_at")
-    .eq("session_token", sessionToken)
-    .single();
-  if (!data || new Date(data.expires_at) < new Date()) return null;
-  return data.professional_id;
-}
+import { requireAdvisorSession } from "@/lib/require-advisor-session";
 
 /**
  * POST /api/advisor-auth/topup
@@ -43,7 +17,7 @@ export async function POST(request: NextRequest) {
     return NextResponse.json({ error: "Too many requests. Please try again later." }, { status: 429 });
   }
 
-  const advisorId = await getAdvisorId(request);
+  const advisorId = await requireAdvisorSession(request);
   if (!advisorId) return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
 
   const body = await request.json().catch(() => ({}));
@@ -145,7 +119,7 @@ export async function POST(request: NextRequest) {
  * Returns the advisor's current credit balance and top-up history.
  */
 export async function GET(request: NextRequest) {
-  const advisorId = await getAdvisorId(request);
+  const advisorId = await requireAdvisorSession(request);
   if (!advisorId) return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
 
   const supabase = createAdminClient();
